==> NoteBook/3.Merge_data.ipynb <==
import pandas as pd
description_embeddings=pd.read_csv("cve_description_embeddings (3).csv")

description_embeddings

description_embeddings.info()

import pandas as pd
cve_data=pd.read_csv("cve_data_cleaned.csv")

cve_data=cve_data.drop('Description', axis=1)
cve_data

cve_data.info()

merged_data = pd.concat([cve_data, description_embeddings], axis=1)
merged_data

merged_data.to_csv('merged_cve_data.csv', index=False)

import pandas as pd
cve_data=pd.read_csv('merged_cve_data.csv')

cve_data

df = pd.read_csv('your_original_dataset.csv')
df.to_parquet('your_dataset.parquet')
